Airflow/requirements.txt
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
comm==0.2.2
debugpy==1.8.5
decorator==5.1.1
executing==2.0.1
h11==0.14.0
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.0.1
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyzmq==26.1.0
requests==2.32.3
selenium==4.23.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
pendulum
jupyterlab
kafka-python
boto3
redis
pymysql